requirements.txt
torch==2.4.1
torchvision==0.19.1
ultralytics==8.3.49
opencv-python>=4.9,<5.0
numpy>=1.24,<2.1
onnxruntime>=1.18,<1.19
pyyaml>=6.0,<7.0
pytest>=7.0
